requirements.txt
fastapi
uvicorn
pydantic
pdfplumber
db-sqlite3
pathlib
watchdog
numpy
pdf2image
pandas
python-multipart
ocrmypdf
SQLAlchemy
python-docx
pytesseract
pyzbar
openpyxl
sentence-transformers
paramiko
pymysql
pypdf
opencv-python
pillow